requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.5.0

# OpenAI
openai>=1.12.0
tiktoken==0.5.2

# Document Processing
pypdf==3.17.0
python-docx==1.1.0

# Vector Store & Search
faiss-cpu==1.7.4
numpy==1.24.3
rank-bm25==0.2.2

# Caching
redis==5.0.1
diskcache==5.6.3

# Async Support
aiofiles==23.2.1
httpx==0.25.2

# Utilities
rich==13.7.0  # Better terminal output
tqdm==4.66.1  # Progress bars